#3116: Use `versionHash` as part of cache-key
When caching taxonomy requests we need to differentiate versions so that
asking for one version doesn't hit the cache of another version.

This solves this by including the `VersionHash` header as part of the
key used when caching taxonomy requests with a header present.

If the value `default` is passed we skip using the header as part of the
key as that is the same as not passing a header at all.