feat: implement graceful shutdowns
This patch introduces graceful shutdowns in the form of a ready flag
that the health endpoints utilizes when determining whether the
readiness probe returns healthy or unhealthy.

When a shutdown signal is detected the ready flag is set to false and a
timeout for actual shutdown starts.

The shutdown process will wait until either all active requests have finished
processing OR the specified timeout is met. Whichever happens first.